Add API to get player history
This is a summary of how a player has done. The current implementation
is not the most efficient, doing multiple queries to get the required
data. This could be improved in a couple of different ways:

1. Use a SQL view to capture the data in a single query
2. Create a `player_histories` table that is updated either via database
triggers or queued background jobs (I didn't want to add redis to this
for the time being)